COMN-20: Take5Dictionary.nextLetters(Take5Match) now returns a match with the expected length: one more than the length of the input.
Note that Take5Dictionary.getStartMatch() has always returned a match of
length 0, as you might expect.

Note also that Take5Dictionary.nextLetters(int, int) continues to return a
match with the wrong length.  But the documentation tells you not to use
this, so you have nothing to complain about.  (I forget now why we didn't
mark this as deprecated...)